requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
distlib==0.3.8
filelock==3.15.4
idna==3.7
pillow==10.3.0
pipenv==2024.0.1
platformdirs==4.2.2
requests==2.32.3
urllib3==2.2.2
virtualenv==20.26.3
